lineralize the execution of events
Before this, when an event listener triggered another event, this new event would be
evaluated and handed through the plugins before the first one is finished. While this
is sensible in cases where the event is supposed to return something, in our case we
want events that are triggered within another event to be executed *after* the current
event is completely finished.

So now, when a command is received and Kabukibot triggers the BOT.CMD event
based on the text message, all other text message handlers will have completed
before the BOT.CMD is run.